domain.types: rename for snake style
